pyproject.toml
[tool.poetry]
name = "tg_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11.7"
aiogram = "^3.3.0"
pydantic = "^2.5.3"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
